core/rawdb: more tests for freezer batching
